requirements.txt
openai==1.58.1
python-dotenv==1.0.1
pillow==11.0.0
